models/mol_moe/requirements.txt
matplotlib==3.9.2
numpy>=1.26.1
pandas>=1.5.3
scikit-learn>=1.5.0
rdkit>=2024.3.5
torch-geometric>=2.3.1
datasets>=2.13.1
huggingface-hub
transformers>=4.38
selfies>=2.1.0
tqdm>=4.66.4
xgboost==2.0.0